requirements.txt
asgiref==3.11.1
Django==6.0.3
django-meta==2.5.1
gunicorn==23.0.0
pillow==12.1.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
